requirements.txt
dash_auth==1.4.1
dash_bootstrap_components==1.3.1
Flask==2.2.2
Flask_Login==0.6.2
plotly==5.13.0
seaborn==0.12.2
gunicorn
pandas
dash-tools
